requirements.txt
# ===============================================================
# CORE & DATA PROCESSING
# ===============================================================
pyspark==3.5.0
numpy>=1.23.0
pandas>=2.0.0
pyarrow>=12.0.0

# ===============================================================
# DATA SOURCES & STORAGE (Kafka, Redis, MinIO, Postgres)
# ===============================================================
kafka-python>=2.0.2
redis>=5.0.0
minio>=7.1.13
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# ===============================================================
# MACHINE LEARNING (LSTM & Analytics)
# ===============================================================
tensorflow>=2.12.0
scikit-learn>=1.1.0
scipy>=1.10.0
joblib>=1.3.0

# ===============================================================
# DASHBOARD & VISUALIZATION
# ===============================================================
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.5.0
seaborn>=0.12.0

# ===============================================================
# UTILS
# ===============================================================
requests>=2.31.0